fix 22/5

+++ app/Http/Controllers/TuongController.php
@@ -36,14 +36,15 @@ class TuongController extends Controller
         }else{
             $check_orders = null;
         }
+        // $check_orders = Order::where('id_user','=',null)->where('phone','=',$new_user->phone)->get();
         $cats = TypeProduct::all();
         $products = Product::where('quantity','>',0)->where('status','=',true)->inRandomOrder()->limit(10)->get();
         $product_hot = Product::where('quantity','>',0)->where('sale','>',0)->where('status','=',true)->inRandomOrder()->limit(3)->get();
         $banners = Banner::all();
         $sliders = Slide::all();
-        // dd($check_orders);
         return view('user.index',compact('products','product_hot','banners','check_orders','cats','sliders'));
     }
+    
     public function get_signUp(){
         $site= "Signup";
         return view('user.pages.Signup.index',compact('site'));
@@ -146,27 +147,27 @@ class TuongController extends Controller
             'url' => route('verify',$new_user->email_verification_token)
         ];
         Mail::to($new_user->email)->send(new VerificationEmail($mailData));
-        $orders = Order::where('id_user','=',null)->where('phone','=',$new_user->phone)->get();
-        $num=0;
-        foreach($orders as $order){
-            $cr_order_code = "USR".$new_user->id_user."_".$num;
-            foreach($order->Cart as $cart){
-                $cart->order_code = $cr_order_code;
-                $cart->id_user = $new_user->id_user;
-                $cart->save();
-            }
-            $order->order_code = $cr_order_code;
-            $order->id_user = $new_user->id_user;
-            $order->save();
-            $num++;
-        }
-        $comment = Comment::where('phone','=',$req["register_phone"])->where('name','=','Guest')->get();
-        foreach($comment as $cmt){
-            $cmt->name=null;
-            $cmt->id_user = $new_user->id_user;
-            $cmt->phone = null;
-            $cmt->save();
-        }
+        // $orders = Order::where('id_user','=',null)->where('phone','=',$new_user->phone)->get();
+        // $num=0;
+        // foreach($orders as $order){
+        //     $cr_order_code = "USR".$new_user->id_user."_".$num;
+        //     foreach($order->Cart as $cart){
+        //         $cart->order_code = $cr_order_code;
+        //         $cart->id_user = $new_user->id_user;
+        //         $cart->save();
+        //     }
+        //     $order->order_code = $cr_order_code;
+        //     $order->id_user = $new_user->id_user;
+        //     $order->save();
+        //     $num++;
+        // }
+        // $comment = Comment::where('phone','=',$req["register_phone"])->where('name','=','Guest')->get();
+        // foreach($comment as $cmt){
+        //     $cmt->name=null;
+        //     $cmt->id_user = $new_user->id_user;
+        //     $cmt->phone = null;
+        //     $cmt->save();
+        // }
         $vali = ["email"=>$new_user->email,"password"=>$req["register_password"]];
         if(Auth::attempt($vali)){
             if(Session::has("cart")){
@@ -228,7 +229,7 @@ class TuongController extends Controller
     public function post_signIn(Request $req){
         $arr_vali = ["email"=>$req["email"],"password"=>$req["password"]];
         if(Auth::attempt($arr_vali)){
-            if(Session::has("cart")){
+            if(Session::has("cart") && Auth::user()->admin == '0'){
                 $cart_session = Session::get("cart");
                 foreach($cart_session as $key => $value){
                     $foundPro = Auth::user()->Cart->where('id_product','=',$value["id_product"])->where('order_code','=',null)->first();
@@ -255,7 +256,6 @@ class TuongController extends Controller
                 }
                 Session::forget("cart");
             };
-
             return redirect('/');
         }else{
             return redirect()->back()->with(["error"=>"Sign in failue. Password or username incorrect"]);
@@ -314,6 +314,16 @@ class TuongController extends Controller
         if(Auth::check()){
             $cart = Cart::where('id_user','=',Auth::user()->id_user)->where('order_code','=',null)->get();
             $address = Auth::user()->Address->sortByDesc('default');
+            $check = true;
+            foreach($address as $add){
+                if($add->default){
+                    $check = false;
+                }
+            };
+            if($check && count($address)>0){
+                $address[0]->default = true;
+                $address[0]->save();
+            }
         }else{
             $cart = Session::get('cart');
             $address = null;
@@ -550,6 +560,7 @@ class TuongController extends Controller
         $order->updated_at = Carbon::now()->format('Y-m-d H:i:s');
         $order->save();
         Session::forget("cart");
+        Session::forget('coupon');
         return redirect('/')->with('order_mess',"Order Successfully, plz wait for Admin Confirm");
     }
     public function admin_cate(){
@@ -590,7 +601,6 @@ class TuongController extends Controller
                 $cart->price = $price;
                 $cart->sale =  $product->sale;
                 $cart->amount = $amount;
-                $cart->created_at = Carbon::now()->format('Y-m-d H:i:s');
                 $cart->save();
             }
         }else if(Session::has("cart")){
@@ -639,7 +649,6 @@ class TuongController extends Controller
                 $cart->sale = $product->sale;
                 $cart->id_product = $id;
                 $cart->amount = $product->quantity < 100 ? $product->quantity:100;
-                $cart->created_at = Carbon::now()->format('Y-m-d H:i:s');
                 $cart->save();
             };
             $num = count(Auth::user()->Cart->where('order_code','=',null))+ 1;
@@ -799,7 +808,7 @@ class TuongController extends Controller
                             <input type='hidden' name='_token' >
                             <input type='hidden' name='max_quan' value='".$cart['max']."'>
                             <div class='col-6 d-flex flex-column justify-content-center align-items-center'>
-                                <div class='input-group input-spinner input-group-sm'>
+                                <div class='input-group input-spinner input-group-sm' data-amount=".$cart["amount"].">
                                     <button type='button' class='btn btn_minus' style='border-radius: 10px 0 0 10px;'  data-field='quantity'>
                                         <i class='bi bi-dash-lg'></i>
                                     </button>
@@ -1041,7 +1050,6 @@ class TuongController extends Controller
         }
         echo $coupon2;
     }
-
     public function get_wishlist(){
         $favourites = Auth::user()->Favourite;
         return view('user.pages.ShopWishlist.index',compact('favourites'));
@@ -1057,23 +1065,25 @@ class TuongController extends Controller
         }else{
             foreach($req->checkFav as $id){
                 $fav = Favourite::find($id);
-                $item = Auth::user()->Cart->where('order_code','=',null)->where('id_product','=',intval($fav->id_product))->first();
-                if($item ){
-                    $item->amount = $item->Product->quantity >($item->amount +100)?$item->amount +100 : ($item->Product->quantity > $item->amount?$item->Product->quantity: $item->amount);
-                    $item->price = $fav->Product->price;
-                    $item->sale = $fav->Product->sale;
-                    $item->updated_at= Carbon::now()->format('Y-m-d H:i:s');
-                    $item->save();
-                }else{
-                    $new_cart = new Cart();
-                    $new_cart->order_code = null;
-                    $new_cart->id_user= Auth::user()->id_user;
-                    $new_cart->id_product = $fav->id_product;
-                    $new_cart->price = $fav->Product->price; 
-                    $new_cart->sale =  $fav->Product->sale; 
-                    $new_cart->amount= $fav->Product->quantity > 100 ? 100 : $fav->Product->quantity;
-                    $new_cart->created_at=Carbon::now()->format('Y-m-d H:i:s');
-                    $new_cart->save();
+                if($fav->Product->status){
+                    $item = Auth::user()->Cart->where('order_code','=',null)->where('id_product','=',intval($fav->id_product))->first();
+                    if($item ){
+                        $item->amount = $item->Product->quantity >($item->amount +100)?$item->amount +100 : ($item->Product->quantity > $item->amount?$item->Product->quantity: $item->amount);
+                        $item->price = $fav->Product->price;
+                        $item->sale = $fav->Product->sale;
+                        $item->updated_at= Carbon::now()->format('Y-m-d H:i:s');
+                        $item->save();
+                    }else{
+                        $new_cart = new Cart();
+                        $new_cart->order_code = null;
+                        $new_cart->id_user= Auth::user()->id_user;
+                        $new_cart->id_product = $fav->id_product;
+                        $new_cart->price = $fav->Product->price; 
+                        $new_cart->sale =  $fav->Product->sale; 
+                        $new_cart->amount= $fav->Product->quantity > 100 ? 100 : $fav->Product->quantity;
+                        $new_cart->created_at=Carbon::now()->format('Y-m-d H:i:s');
+                        $new_cart->save();
+                    }
                 }
             }
             return redirect('/order');
@@ -1267,9 +1277,6 @@ class TuongController extends Controller
         $address->default = true;
         $address->save();
         return redirect()->back()->with('message','Change Default Address successfully');
-    }
-    public function get_payment(){
-        return view('user.pages.About.payment');
     }
     public function get_feedback($code){
         $order = Order::where('order_code','=',$code)->first();
@@ -1635,7 +1642,10 @@ class TuongController extends Controller
         $new_message->unread_mess = $unread_mess;
         echo $new_message;
     }
-    public function clear_grchat($code){
+    public function clear_grchat($code =null){
+        if(!$code){
+            return redirect()->back();
+        }
         $group = Groupmessage::where('code_group','=',$code)->first();
         foreach($group->Message as $mess){
             $mess->delete();
@@ -1664,6 +1674,7 @@ class TuongController extends Controller
     }
     public function model_coupon($code){
         $coupon = Coupon::where('code','=',$code)->where('status',true)->first();
+        $coupon->used = count(Order::where('id_user','=',Auth::user()->id_user)->where('code_coupon','=',$code)->get());
         if($coupon){
             echo $coupon;
         }else{
@@ -1723,7 +1734,7 @@ class TuongController extends Controller
     }
     public function create_newpassword(Request $req){
         $user = User::where('email','=',$req['account_email'])->first();
-        $user->password = bcrypt($req['new_password']);
+        $user->password = bcrypt($req['new_password2']);
         $user->updated_at = Carbon::now()->format('Y-m-d H:i:s');
         $user->save();
         $find_email = ResetPassword::find($req['account_email'])->first();
